ipcache: Implement Shutdown
[backporter's note: backed out of propagating the context to the
controller, shutdown func of the trigger removed

this means that we don't wait for shutdown to actually complete, but
that's just best effort.]

Implement a Shutdown() function for the IPCache which shuts down the
goroutines associated with the IPcache and ensures they are completed
before returning. This way, other components should not continue to
cause the IPCache to asynchronously continue to execute work. This
should hopefully fix issues in unit tests where the IPCache defers some
work that interacts with the identityAllocator after it has been cleaned
up.

Suggested-by: Aditi Ghag <aditi@cilium.io>
Signed-off-by: Joe Stringer <joe@cilium.io>
Signed-off-by: David Bimmler <david.bimmler@isovalent.com>